Added an Activity Indicator to show that the meal page is loading
It seems that the first time in a day that I open the app, it takes a while to fetch the information from spoonacular. I don't know if it is loading or just stuck cause I close the app and try like another two times before it works. The activity indicator will show that it is loading.