Popup Displays Ingrediants and Insructions
In my meals screen.js above the first fetch I added a list of api keys. When app says recipes.map is not a function it means the limit for the day on the api key has been reached. Now we can easily switch. The api fetch are as follows: One in mymealsscreen, two in mealdetails