Updates with positioning. Scrollview fixed. Modal fixed
Search bar is now outside of the scroll view so it is always available. Constant modalstyles removed, all modal stuff is now with the rest of styles. Modal also changed to correct size and color. Maybe too large for its purpose at this time. Colors for button on purple changed to cyan.