Changed some sizes. Limited recipe titles to 2 lines.
Modal height, and recipe box heights changed. There was an issue with three line title of recipes not fitting in the boxes, so I limited it to 2 lines. The modal shows the full recipe title anyways. Lengthy recipe titles also took up a lot of the space from the modal, so I increased the height to give instructions and ingredients more space.